+++ intervention-process.php
@@ -15,10 +15,10 @@ $logged_level = $_SESSION['logged_level'];
 unset($err_msg);
 
 $description        = param_post('description');
-$equipment_id       = param_post('equipment_id');
-$supplier_id        = param_post('supplier_id');
+$equipment_id       = (int)param_post('equipment_id');
+$supplier_id        = (int)param_post('supplier_id');
 $date               = param_post('date');
-$intervention_id    = param_post('intervention_id');
+$intervention_id    = (int)param_post('intervention_id');
 
 
 $recipe = '';
